requirements.txt
python == 3.10.4
Django == 2.2
djangorestframework == 3.13.1
django-rest-swagger == 2.2.0